+++ README.md
@@ -1,6 +1,6 @@
 # skrebets-sprint11.github.io
 
-*version 0.0.1*
+*version 0.0.2*
 
 ## Link
